+++ scripts/check_guard_name_style.sh
@@ -4,46 +4,47 @@ function CheckGuardNameStyle () {
   filename=$1
 
   # Get the guard name
-  guardname="CARTA_"
+  guard_name="CARTA_"
 
-  names=(${filename//\// })
-  for i in "${names[@]}"
+  # Split the file path name by '/' as an array
+  sub_names=(${filename//\// })
+  for i in "${sub_names[@]}"
   do
-    guardname+=$i
-    guardname+="_"
+    guard_name+=$i
+    guard_name+="_"
   done
 
-  guardname=$(echo $guardname | tr '[:lower:]' '[:upper:]')
-  guardname=$(echo $guardname | sed "s/\./\_/g")
+  guard_name=$(echo "${guard_name}" | tr '[:lower:]' '[:upper:]')
+  guard_name=$(echo "${guard_name}" | sed "s/\./\_/g")
 
-  # Get the line number for the first occurence of '#ifndef ...'
+  # Get the line number for the first occurrence of '#ifndef ...'
   line_num=$(grep -m1 -n -- '^#ifndef' "${filename}")
   line_num="${line_num%%:*}"
 
   if [[ $line_num -gt 0 ]]; then
-    # Replace with the guard name in the line number for the first occurence of '#ifndef ...'
-    if [[ "$OSTYPE" == "darwin"* ]]; then
-      sed -i '' "${line_num}s/.*/#ifndef $guardname/" $filename
+    # Replace with the guard name in the line number for the first occurrence of '#ifndef ...'
+    if [[ "${OSTYPE}" == "darwin"* ]]; then
+      sed -i '' "${line_num}s/.*/#ifndef ${guard_name}/" "${filename}"
     else
-      sed -i "${line_num}s/.*/#ifndef $guardname/" $filename
+      sed -i "${line_num}s/.*/#ifndef ${guard_name}/" "${filename}"
     fi
 
-    # Get the line number after the first occurence of '#ifndef ...'
+    # Get the line number after the first occurrence of '#ifndef ...'
     line_num=$((line_num + 1))
 
-    # Replace with the guard name in the line number after the first occurence of '#ifndef ...'
-    if [[ "$OSTYPE" == "darwin"* ]]; then
-      sed -i '' "${line_num}s/.*/#define $guardname/" $filename
+    # Replace with the guard name in the line number after the first occurrence of '#ifndef ...'
+    if [[ "${OSTYPE}" == "darwin"* ]]; then
+      sed -i '' "${line_num}s/.*/#define ${guard_name}/" "${filename}"
     else
-      sed -i "${line_num}s/.*/#define $guardname/" $filename
+      sed -i "${line_num}s/.*/#define ${guard_name}/" "${filename}"
     fi
   fi
 
-  # Replace the last occurence of matched '#endif ...'
-  if [[ "$OSTYPE" == "darwin"* ]]; then
-    sed -i '' "$ s/^#endif .*$/#endif \/\/ $guardname/" $filename
+  # Replace the last occurrence of matched '#endif ...'
+  if [[ "${OSTYPE}" == "darwin"* ]]; then
+    sed -i '' "$ s/^#endif .*$/#endif \/\/ ${guard_name}/" "${filename}"
   else
-    sed -i "$ s/^#endif .*$/#endif \/\/ $guardname/" $filename
+    sed -i "$ s/^#endif .*$/#endif \/\/ ${guard_name}/" "${filename}"
   fi
 }
 
@@ -51,13 +52,13 @@ function CheckFolder() {
   # Get file names
   folder=$1
   suffix=$2
-  filenames=( $(find $folder -name $suffix) )
+  filenames=($(find "${folder}" -name "${suffix}"))
 
   # Replace the guard names for each of the files
   for j in "${filenames[@]}"
   do
     filename=$j
-    CheckGuardNameStyle $filename
+    CheckGuardNameStyle "${filename}"
   done
 }
 
@@ -70,4 +71,3 @@ CheckFolder "src" "*.tcc"
 
 CheckFolder "test" "*.h"
 CheckFolder "test" "*.tcc"
-